Fin de modification.

+++ src/bankAccount/BankAccount.java
@@ -11,11 +11,20 @@ public class BankAccount {
 	protected int amount;
 	protected Inhabitant owner;
 	
+	/**
+	 * First constructor of the class BankAccount. this one take only one parameter.
+	 * @param owner
+	 */
 	public BankAccount(Inhabitant owner) {
 		this.owner = owner;
 		this.amount = 5000;
 	}
 
+	/**
+	 * Second constructor of the BankAccount. this one take two parameter.
+	 * @param amount
+	 * @param owner
+	 */
 	public BankAccount(int amount, Inhabitant owner) {
 		this.amount = amount;
 		this.owner = owner;
@@ -25,7 +34,7 @@ public class BankAccount {
 	 * function which debits an amount from a bank account.
 	 * @param amount : the amount to remove.
 	 */
-	public void debit(double amount) {
+	public void debit(int amount) {
 		this.amount -= amount;
 		System.out.println(" - "+amount+" are debited from "+owner.getName()+" account whose balance is now "+this.amount+ " euros");
 	}
@@ -34,15 +43,22 @@ public class BankAccount {
 	 * function which credit an amount to a bank account.
 	 * @param amount : the amount to add.
 	 */
-	public void credit(double amount) {
+	public void credit(int amount) {
 		this.amount += amount;
 		System.out.println(" + "+owner.getName()+" account is credited with "+amount+" euros; its balance is now "+this.amount+ " euros");
 	}
 	
+	/**
+	 * function which return the amount 
+	 * @return amount
+	 */
 	public int getAmount() {
 		return amount;
 	}
-	
+	/**
+	 * function which return the Owner
+	 * @return owner
+	 */
 	public Inhabitant getOwner() {
 		return owner;
 	}
